Services implemented

+++ src/main/java/com/example/springPostgres/IService/IAnagraficaService.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 public interface IAnagraficaService {
 
-    public Anagrafica newAnagrafica(int id, String nome, String cognome);
+    public Anagrafica newAnagrafica(Long id, String nome, String cognome);
     public boolean deleteAnagrafica(Long id);
     public Optional<Anagrafica> updateAnagrafica(Long id, String nome, String cognome);
     public Optional<Anagrafica> anagraficaById(Long id);
